move systemBuilder to packages/core/ and refactor it into separate files
