+++ app/js/sim/workers/systemBuilder.worker.js
@@ -1,4 +1,4 @@
-import { System, GridSystem, SoPhysics } from '-/sim/systemBuilder'
+import System from '@void/core/system-builder/System'
 
 self.onmessage = function (e) {
   const system = new System(1, 1, e.data[0], e.data[1], e.data[2])
