+++ packages/core/system-builder/config.js
@@ -1,0 +1,6 @@
+const config = {
+  G: 2.93558 * Math.pow(10, -4),
+  epsilon: 0.01
+}
+
+export default config
