+++ packages/core/system-builder/soPhysics.js
@@ -1,225 +1,5 @@
-import uuid from 'uuid/v4'
-// import * as zeta from  'riemann-zeta'
-import { randomUniform, guid } from '-/utils'
-
-class Body {
-  constructor (body_data = []) {
-    if (body_data == []) {
-      body_data = [ 'body', 1, 0, 0, 0, 0, 0, 0, 0, 0, 0 ]
-    }
-    this.name = body_data[0]
-    this.mass = body_data[1]
-    this.position = new Point([ body_data[2], body_data[3], body_data[4] ])
-    this.velocity = new Point([ body_data[5], body_data[6], body_data[7] ])
-    this.acceleration = new Point([ body_data[8], body_data[9], body_data[10] ])
-    this.orientation = new Point()
-    this.angVelocity = new Point()
-    this.acceleration.reset()
-    this.radius = (0.018 * Math.pow(this.mass, 0.35)) / 149600000000
-  }
-}
-class Point {
-  constructor (position_data = [ 0, 0, 0 ]) {
-    this.x = position_data[0]
-    this.y = position_data[1]
-    this.z = position_data[2]
-  }
-  reset () {
-    this.x = 0
-    this.y = 0
-    this.z = 0
-  }
-}
-
-class Star {
-  constructor (
-    xPos = 30000,
-    yPos = 0,
-    zPos = 0,
-    player = 'duh',
-    aName = 'default'
-  ) {
-    this.body = new Body([ 'body', 1, xPos, yPos, zPos, 0, 0, 0, 0, 0, 0 ])
-    this.color = [ 0.0, 0.0, 1.0 ]
-    this.mass = 16
-    this.radius = 6.6
-    this.temp = 33000
-    this.buildrandom()
-  }
-  buildrandom () {
-    let starrand = Math.random() * 10000
-    if (starrand < 7600) {
-      this.mtype()
-    }
-    if (starrand < 8800 && starrand > 7600) {
-      this.ktype()
-    }
-    if (starrand < 9400 && starrand > 8800) {
-      this.gtype()
-    }
-    if (starrand < 9700 && starrand > 9400) {
-      this.ftype()
-    }
-    if (starrand < 9800 && starrand > 9900) {
-      this.atype()
-    }
-    if (starrand < 9900 && starrand > 9950) {
-      this.btype()
-    }
-    if (starrand < 9950) {
-      this.otype()
-    }
-  }
-  otype () {
-    this.color = [ 1.0, 0.0, 0.0 ]
-    this.mass = 16
-    this.radius = 6.6
-    this.temp = 33000
-  }
-  btype () {
-    this.color = [ 0.5, 0.5, 0.8, 1.0 ]
-    this.mass = 2.1
-    this.radius = 1.8
-    this.temp = 10000
-  }
-  atype () {
-    this.color = [ 1.0, 1.0, 1.0, 1.0 ]
-    this.mass = 1.4
-    this.radius = 1.4
-    this.temp = 7500
-  }
-  ftype () {
-    this.color = [ 1.0, 1.0, 0.8, 1.0 ]
-    this.color = [ 1.0, 0.0, 0.0 ]
-    this.mass = 1
-    this.radius = 1
-    this.temp = 6000
-  }
-  gtype () {
-    this.color = [ 1.0, 1.0, 0.0, 1.0 ]
-    this.mass = 0.8
-    this.radius = 0.9
-    this.temp = 5200
-  }
-  ktype () {
-    this.color = [ 1.0, 0.2, 0.2, 1.0 ]
-    this.mass = 0.7
-    this.radius = 0.5
-    this.temp = 3700
-  }
-  mtype () {
-    this.color = [ 1.0, 0.0, 0.0, 1.0 ]
-    this.mass = 0.2
-    this.radius = 0.2
-    this.temp = 2000
-  }
-}
-
-class GridSystem {
-  constructor (bodies = []) {
-    this.count = bodies.length
-    this.player = 0
-    this.names = [ '' ]
-    this.mass = [ 0.0 ]
-    this.rad = [ 0.0 ]
-    this.pos = [ [ 0.0, 0.0, 0.0 ] ]
-    this.ori = [ [ 0.0, 0.0, 0.0 ] ]
-    this.vel = [ [ 0.0, 0.0, 0.0 ] ]
-    this.acc = [ [ 0.0, 0.0, 0.0 ] ]
-    this.allocated = 1
-    this.collisions = []
-    this.removed = []
-    for (let i = 0; i < this.count; i++) {
-      this.addSpace()
-    }
-    let i = 0
-    for (let body of bodies) {
-      if (body.name == 'player') {
-        this.player = i
-      }
-      this.insertBody(body, i)
-      i += 1
-    }
-    // Void.log.debug(bodies)
-  }
-  getPlayerIndex () {
-    let i = 0
-    let lasti = 0
-    for (let name of this.names) {
-      if (name == 'player') {
-        this.player = i
-        lasti = i
-        return i
-      } else {
-        i += 1
-      }
-    }
-    return i
-  }
-
-  insertBody (body, i) {
-    this.names[i] = body.name
-    this.mass[i] = body.mass
-    this.rad[i] = body.radius
-    this.pos[i][0] = body.position.x
-    this.pos[i][1] = body.position.y
-    this.pos[i][2] = body.position.z
-
-    this.ori[i][0] = body.orientation.x
-    this.ori[i][1] = body.orientation.y
-    this.ori[i][2] = body.orientation.z
-
-    this.vel[i][0] = body.velocity.x
-    this.vel[i][1] = body.velocity.y
-    this.vel[i][2] = body.velocity.z
-
-    this.acc[i][0] = body.acceleration.x
-    this.acc[i][1] = body.acceleration.y
-    this.acc[i][2] = body.acceleration.z
-  }
-
-  moveBody (source, dest) {
-    this.names[dest] = this.names[source]
-    this.mass[dest] = this.mass[source]
-    this.rad[dest] = this.rad[source]
-    this.pos[dest] = this.pos[source]
-    this.ori[dest] = this.ori[source]
-    this.vel[dest] = this.vel[source]
-    this.acc[dest] = this.acc[source]
-    this.names[source] = 'OLD'
-  }
-
-  removeBody (i) {
-    if (i != this.player) {
-      if (i == this.count - 1) {
-        let foo = 1
-      } else {
-        this.moveBody(this.count - 1, i)
-      }
-      this.count -= 1
-      this.getPlayerIndex()
-    }
-  }
-
-  resetAcc () {
-    for (let i = 0; i < this.count; i++) {
-      this.acc[i] = [ 0.0, 0.0, 0.0 ]
-    }
-  }
-
-  addSpace () {
-    this.allocated += 1
-    this.names.push('')
-    this.mass.push(0.0)
-    this.rad.push(0.0)
-    this.pos.push([ 0.0, 0.0, 0.0 ])
-    this.ori.push([ 0.0, 0.0, 0.0 ])
-    this.vel.push([ 0.0, 0.0, 0.0 ])
-    this.acc.push([ 0.0, 0.0, 0.0 ])
-  }
-}
-const G = 2.93558 * Math.pow(10, -4)
-const epsilon = 0.01
+import GridSystem from './GridSystem'
+import { evaluate } from './utils'
 
 class soPhysics {
   constructor (
@@ -809,344 +589,4 @@ class soPhysics {
   }
 }
 
-class System {
-  constructor (
-    seed = 1,
-    starcount = 1,
-    bodycount = 1,
-    abodyDistance = 2,
-    abodySpeed = 0.05
-  ) {
-    this.seed = seed
-    this.star = new Star()
-    this.starCount = starcount
-    this.bodyCount = bodycount
-    if (bodycount > 1) {
-      var here = 'pewp'
-    }
-    this.bodies = []
-    this.bodyDistance = abodyDistance
-    this.bodySpeed = abodySpeed
-    if (seed != 0) {
-      this.build()
-    } else {
-      this.buildSol()
-    }
-    // Void.log.debug('bodyCount: ' + this.bodies.length)
-    this.stability = 0.5 - evaluate(this.bodies)
-
-    this.avgStability = 0.5 - evaluate(this.bodies)
-  }
-  moveToStar () {
-    for (let body of this.bodies) {
-      body.position.x += this.star.body.position.x
-      body.position.y += this.star.body.position.y
-      body.position.z += this.star.body.position.z
-    }
-  }
-  getStar (body_data) {
-    body_data.push(randomUniform(0.6, 4))
-    for (let j = 0; j <= 2; j++) {
-      body_data.push(0.0)
-    }
-    body_data.push(0.0)
-    for (let j = 0; j <= 2; j++) {
-      body_data.push(0.0)
-    }
-    body_data.push(0.0)
-    return body_data
-  }
-  // getSymPlanets () {
-  //   let body_data = []
-  //
-  //   // body name
-  //   body_data.push(`body ${uuid()}`)
-  //   body_data.push(randomUniform(0.0, 0.1))
-  //   if (quadrantconst > 0) {
-  //     body_data.push(randomUniform(0, this.bodyDistance))
-  //     body_data.push(randomUniform(0, this.bodyDistance))
-  //     body_data.push(0.0)
-  //     body_data.push(randomUniform(0, this.bodySpeed))
-  //     body_data.push(randomUniform(-this.bodySpeed, 0))
-  //     body_data.push(0.0)
-  //   }
-  //   if (quadrantconst < 0) {
-  //     body_data.push(randomUniform(-this.bodyDistance, 0))
-  //     body_data.push(randomUniform(-this.bodyDistance, 0))
-  //     body_data.push(0.0)
-  //     body_data.push(randomUniform(-this.bodySpeed, 0))
-  //     body_data.push(randomUniform(0, this.bodySpeed))
-  //     body_data.push(0.0)
-  //   }
-  // }
-  getDirectedPlanet (quadrantconst = 1) {
-    let body_data = []
-
-    // body name
-    body_data.push(`body ${uuid()}`)
-    let min = 8 * Math.pow(10, -10)
-    let max = 0.00001
-
-    let test = Math.random()
-    if (test >= 0.8) {
-      min = 0.00001
-      max = 0.01
-    }
-    let mass = randomUniform(min, max)
-
-    body_data.push(mass)
-    if (quadrantconst > 0) {
-      body_data.push(randomUniform(0, this.bodyDistance))
-      body_data.push(randomUniform(0, this.bodyDistance))
-      body_data.push(randomUniform(0, this.bodyDistance / 128))
-    }
-    if (quadrantconst < 0) {
-      body_data.push(randomUniform(-this.bodyDistance, 0))
-      body_data.push(randomUniform(-this.bodyDistance, 0))
-      body_data.push(randomUniform(-this.bodyDistance / 128))
-    }
-    if (quadrantconst > 0) {
-      body_data.push(randomUniform(0, this.bodySpeed))
-      body_data.push(randomUniform(-this.bodySpeed, 0))
-      body_data.push(randomUniform(0, this.bodySpeed / 128))
-    }
-    if (quadrantconst < 0) {
-      body_data.push(randomUniform(-this.bodySpeed, 0))
-      body_data.push(randomUniform(0, this.bodySpeed))
-      body_data.push(randomUniform(-this.bodySpeed / 128))
-    }
-    return body_data
-  }
-
-  getPlanet (body_data) {
-    body_data.push(randomUniform(0.00000000001, 0.01))
-    for (let j of range(0, 2)) {
-      body_data.push(randomUniform(-this.bodyDistance, this.bodyDistance))
-    }
-    body_data.push(0.0)
-    for (let j of range(0, 2)) {
-      body_data.push(randomUniform(-this.bodySpeed, this.bodySpeed))
-    }
-    body_data.push(0.0)
-    return body_data
-  }
-  buildSol () {
-    this.bodies = []
-    let body_data = [ 'Sol', 1, 0, 0, 0, 0, 0, 0, 0, 0, 0 ]
-    this.bodies.push(Body(body_data))
-    body_data = [ 'Earth', 0.000003, 0, 1, 0, 0.04, 0, 0, 0, 0, 0 ]
-    this.bodies.push(Body(body_data))
-  }
-  build () {
-    for (let i = 0; i < this.bodyCount; i++) {
-      if (i < this.starCount) {
-        this.addStar()
-      } else {
-        this.addSinglePlanet()
-      }
-    }
-  }
-
-  addStar () {
-    let body_data = this.getStar([ 'star' ])
-    let body = new Body(body_data)
-    this.bodies.push(body)
-  }
-  reverseBody (adata) {
-    let bdata = adata
-    bdata[2] = 0 - adata[2]
-    bdata[3] = 0 - adata[3]
-    bdata[4] = 0 - adata[4]
-    bdata[5] = 0 - adata[5]
-    bdata[6] = 0 - adata[6]
-    bdata[7] = 0 - adata[7]
-    return bdata
-  }
-  rotateBody (adata) {
-    let bdata = adata
-    bdata[2] = adata[3]
-    bdata[3] = 0 - adata[2]
-    bdata[4] = 0 // adata[4]
-
-    bdata[5] = adata[6]
-    bdata[6] = 0 - adata[5]
-    bdata[7] = adata[7]
-    return bdata
-  }
-
-  invertXbody (adata) {
-    let bdata = [] // adata
-    bdata[0] = `body ${uuid()}`
-    bdata[1] = adata[1]
-    bdata[2] = -1 * adata[2]
-    bdata[3] = adata[3]
-    bdata[4] = adata[4]
-
-    bdata[5] = adata[5]
-    bdata[6] = -1 * adata[6]
-    bdata[7] = adata[7]
-    return bdata
-  }
-  invertYbody (adata) {
-    let bdata = [] // adata
-    bdata[0] = `body ${uuid()}`
-    bdata[1] = adata[1]
-    bdata[2] = adata[2]
-    bdata[3] = -1 * adata[3]
-    bdata[4] = adata[4]
-
-    bdata[5] = -1 * adata[5]
-    bdata[6] = adata[6]
-    bdata[7] = adata[7]
-    return bdata
-  }
-
-  addSinglePlanet () {
-    // Void.log.debug('adding Body')
-    let body_data = this.getDirectedPlanet()
-    let aBody = new Body(body_data)
-    let bBody = new Body(this.reverseBody(body_data))
-    let cBody = new Body(this.invertYbody(body_data))
-    let dBody = new Body(this.invertXbody(body_data))
-    let otherBodies = []
-    otherBodies.push(this.bodies[0])
-    otherBodies.push(aBody)
-    otherBodies.push(bBody)
-    otherBodies.push(cBody)
-    otherBodies.push(dBody)
-    let fitness = this.evaluateN(otherBodies)
-
-    while (fitness < 0.1 || fitness > 1) {
-      // Void.log.debug('testing configuration')
-      body_data = this.getDirectedPlanet()
-      aBody = new Body(body_data)
-      bBody = new Body(this.reverseBody(body_data))
-      cBody = new Body(this.invertYbody(body_data))
-      dBody = new Body(this.invertXbody(body_data))
-
-      let otherBodies = []
-      otherBodies.push(this.bodies[0])
-      otherBodies.push(aBody)
-      otherBodies.push(bBody)
-      otherBodies.push(cBody)
-      otherBodies.push(dBody)
-      fitness = this.evaluateN(otherBodies)
-    }
-    this.bodies.push(aBody)
-    this.bodies.push(bBody)
-    this.bodies.push(cBody)
-    this.bodies.push(dBody)
-    return aBody
-  }
-  addPlanet () {
-    // Void.log.debug('adding body')
-    let body_data = []
-    body_data.push(guid())
-    body_data = this.getPlanet(body_data)
-    let aBody = new Body(body_data)
-    this.bodies.push(aBody)
-    // Void.log.debug('new stability')
-    // Void.log.debug(this.evaluate(this.bodies))
-    while (this.evaluate(this.bodies) > 1) {
-      this.bodies.pop()
-      this.addPlanet()
-    }
-  }
-
-  //   evaluateBodies( someBodies) {
-  //           kinetic=0.0;
-  //           potential=0.0;
-  //           G=2.93558*10**-4;
-  //           for (body of someBodies) {
-  //                   vel = body.velocity;
-  //                   vel_sq = (vel.x**2 + vel.y**2 + vel.z**2);
-  //                   kinetic += 0.5*body.mass*vel_sq;
-  //                 }
-  //           for (i of range(0,len(someBodies))) {
-  //                   current_body=someBodies[i];
-  //                   current_position=current_body.position;
-  //                   for (j of range(0,i)) {
-  //                           other_body=someBodies[j];
-  //                           other_position=other_body.position;
-  //                           d_x=(other_position.x-current_position.x);
-  //                           d_y=(other_position.y-current_position.y);
-  //                           d_z=(other_position.z-current_position.z);
-  //                           radius = (d_x**2 + d_y**2 + d_z**2)**(0.5);
-  //                           if (radius >0 ) {
-  //                                   potential -= G*current_body.mass*other_body.mass/radius;}
-  //                                 }
-  //           }
-  //           try {
-  //                   return abs(kinetic/potential);
-  //           } catch () {
-  //                   return 100.0;
-  //                 }
-  // }
-
-  evaluateN (somebodies) {
-    let tempSys = new System()
-    tempSys.bodies = somebodies
-    let tempEval = new soPhysics(tempSys, 1000000, 0.01)
-    return tempEval.sumFit
-  }
-
-  bodies () {
-    return this.bodies
-  }
-}
-
-const evaluate = someBodies => {
-  // Void.log.debug('bodies')
-  // Void.log.debug(someBodies)
-  let kinetic = 0.0
-  let potential = 0.0
-  let G = 2.93558 * Math.pow(10, -4)
-  for (let body of someBodies) {
-    // Void.log.debug(body)
-    let vel = body.velocity
-    let vel_sq = Math.pow(vel.x, 2) + Math.pow(vel.y, 2) + Math.pow(vel.z, 2)
-    kinetic += 0.5 * body.mass * vel_sq
-  }
-  for (let i = 0; i < someBodies.length; i++) {
-    let current_body = someBodies[i]
-    let current_position = current_body.position
-    for (let j = 0; j < i; j++) {
-      let other_body = someBodies[j]
-      let other_position = other_body.position
-      let d_x = other_position.x - current_position.x
-      let d_y = other_position.y - current_position.y
-      let d_z = other_position.z - current_position.z
-      let radius = Math.pow(
-        Math.pow(d_x, 2) + Math.pow(d_y, 2) + Math.pow(d_z, 2),
-        0.5
-      )
-      if (radius > 0) {
-        potential -= (G * current_body.mass * other_body.mass) / radius
-      }
-    }
-  }
-  try {
-    return Math.abs(kinetic / potential)
-  } catch (err) {
-    return 100
-  }
-}
-
-const convertSystemToMeters = system => {
-  return system.bodies.map(body => {
-    body.position.x *= 149600000000
-    body.position.y *= 149600000000
-    body.position.z *= 149600000000
-
-    body.velocity.x *= 149600000000
-    body.velocity.y *= 149600000000
-    body.velocity.z *= 149600000000
-
-    //    body.radius *= 149600000000
-
-    body.radius = 0.018 * Math.pow(body.mass * 2 * Math.pow(10, 30), 0.35)
-    return body
-  })
-}
-
-export { System, GridSystem, soPhysics, convertSystemToMeters, evaluate, Point }
+export default soPhysics
